pyproject.toml
[project]
name = "puzzle-game"
version = "0.1.0"
requires-python = ">=3.12.0"
dependencies = [
    "pygame-ce>=2.5.6",
]
